packages/scheml/python/requirements.txt
numpy==1.26.4
scikit-learn==1.5.2
skl2onnx==1.16.0
onnx==1.16.0
flaml==2.3.0
joblib==1.4.2
faiss-cpu==1.8.0
